% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{asCanonical}
\alias{asCanonical}
\title{Coerce a Vector of Cluster Labels to Canonical Form}
\usage{
asCanonical(partition)
}
\arguments{
\item{partition}{A numeric vector representing a set partition of the
integers \eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
A numeric vector representing \code{partition}, but now in canonical
  form.
}
\description{
Coerce a Vector of Cluster Labels to Canonical Form
}
