% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{asSetPartition}
\alias{asSetPartition}
\title{Coerce a Set Partition as Numeric Vectors of Cluster Labels to a List
Structure}
\usage{
asSetPartition(partition)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
The set partition in a list structure.
}
\description{
Coerce a Set Partition as Numeric Vectors of Cluster Labels to a List
Structure
}
