% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{joinExistingCluster}
\alias{joinExistingCluster}
\title{Join Item to an Existing Cluster}
\usage{
joinExistingCluster(i, join, partition)
}
\arguments{
\item{i}{Item index as an integer vector of length one}

\item{join}{Label or index of cluster that \code{i} must join}

\item{partition}{Set partition of the integers \eqn{1}, ..., \eqn{n}
represented as either a numeric vector of cluster labels, or a list
containing subsets of these integers}
}
\value{
Updated partition.
}
\description{
Join Item to an Existing Cluster
}
