% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{nClusters}
\alias{nClusters}
\title{Count the Number of Clusters in a Set Partition}
\usage{
nClusters(partition)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
The number of clusters in the given set partition as a numeric vector
  of length one.
}
\description{
Count the Number of Clusters in a Set Partition
}
\examples{

p <- c(0,1,1,2,3,2,4,4,2)
nClusters(p)
}
