% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p18_bern}
\alias{p18_bern}
\alias{p18_bern_1}
\alias{p18_bern_2}
\alias{p18_bern_3}
\title{Multivariate Independent Bernoulli Data (p = 18)}
\format{
An object of class \code{data.frame} with 100 rows and 18 columns.

An object of class \code{data.frame} with 100 rows and 18 columns.

An object of class \code{data.frame} with 100 rows and 18 columns.
}
\usage{
p18_bern_1

p18_bern_2

p18_bern_3
}
\description{
Multivariate categorical data, generated as five clusters of independent variables with different
  success probabilities.
}
\keyword{datasets}
