% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p18_bernoulli}
\alias{p18_bernoulli}
\title{Multivariate Independent Bernoulli Data (p = 18)}
\format{
A data frame with 100 rows and 18 variables.
}
\usage{
p18_bernoulli
}
\description{
Multivariate categorical data, generated as five clusters of independent variables with different
  success probabilities.
}
\keyword{datasets}
