% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p18_corr_mvn}
\alias{p18_corr_mvn}
\title{Correlated Multivariate Normal Data (p = 18)}
\format{
A data frame with 100 rows and 18 variables
}
\usage{
p18_corr_mvn
}
\description{
Multivariate continuous data, generated as five clusters with an AR(1) correlation structure
  determining the common covariance matrix for each observation.
}
\keyword{datasets}
