% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p18_mvn}
\alias{p18_mvn}
\title{Independent Multivariate Normal Data (p = 18)}
\format{
A data frame with 100 rows and 18 variables
}
\usage{
p18_mvn
}
\description{
Multivariate continuous data, generated as five clusters of independent variables with different
  means.
}
\keyword{datasets}
