% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p6_bernoulli}
\alias{p6_bernoulli}
\title{Multivariate Independent Bernoulli Data (p = 6)}
\format{
A data frame with 100 rows and 6 variables
}
\usage{
p6_bernoulli
}
\description{
Multivariate categorical data, generated as five clusters of independent variables with different
  success probabilities.
}
\keyword{datasets}
