% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p6_big_bern}
\alias{p6_big_bern}
\alias{p6_big_bern_1}
\alias{p6_big_bern_2}
\alias{p6_big_bern_3}
\title{Large Sample Multivariate Independent Bernoulli Data (p = 6)}
\format{
An object of class \code{data.frame} with 1000 rows and 6 columns.

An object of class \code{data.frame} with 1000 rows and 6 columns.

An object of class \code{data.frame} with 1000 rows and 6 columns.
}
\usage{
p6_big_bern_1

p6_big_bern_2

p6_big_bern_3
}
\description{
Multivariate categorical data, generated as five clusters of independent variables with different
  success probabilities.
}
\keyword{datasets}
