% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p6_mvn}
\alias{p6_mvn}
\title{Independent Multivariate Normal Data (p = 6)}
\format{
A data frame with 100 rows and 6 variables.
}
\usage{
p6_mvn
}
\description{
Multivariate continuous data, generated as five clusters of independent variables with different
  means.
}
\keyword{datasets}
