% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sizeOfLargestCluster}
\alias{sizeOfLargestCluster}
\title{Calculate the Number of Items in the Largest Cluster of a Set Partition}
\usage{
sizeOfLargestCluster(partition)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
The number of items in the largest cluster of the given partition as
  a numeric vector of length one.
}
\description{
Calculate the Number of Items in the Largest Cluster of a Set Partition
}
\examples{

p <- c(0,1,1,1,1,1,2)
sizeOfLargestCluster(p)
}
