% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getThetas}
\alias{getThetas}
\title{Get theta Parameters from a Numeric Vector of Cluster Labels and Unique phi
Values}
\usage{
getThetas(partition, phi)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}

\item{phi}{A list of unique model parameters whose length must equal the
number of unique cluster labels in \code{partition}}
}
\value{
A numeric vector of model parameters \eqn{theta_1}, ...,
  \eqn{theta_n}.
}
\description{
Get theta Parameters from a Numeric Vector of Cluster Labels and Unique phi
Values
}
