% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{isCanonical}
\alias{isCanonical}
\title{Check if a Vector of Cluster Labels is in Canonical Form}
\usage{
isCanonical(partition)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
Logical, indicating whether \code{partition} is in canonical form.
}
\description{
Check if a Vector of Cluster Labels is in Canonical Form
}
