% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{partitionEntropy}
\alias{partitionEntropy}
\title{Calculate the Entropy of a Set Partition}
\usage{
partitionEntropy(partition)
}
\arguments{
\item{partition}{A numeric vector representing a partition of the integers
\eqn{1}, ..., \eqn{n} using cluster labels}
}
\value{
Calculated partition entropy as a numeric vector of length one
}
\description{
Calculate the Entropy of a Set Partition
}
\examples{

p <- c(0,0,0,1,1,2) # n = 6, 3 unique clusters
partitionEntropy(p)
}
